+++ test/test_backends.py
@@ -108,9 +108,24 @@ def test_docker_encoding(host):
     ('u:P@h', HostSpec('h', None, 'u', 'P')),
     ('u@h', HostSpec('h', None, 'u', None)),
     ('h', HostSpec('h', None, None, None)),
+    ('pr%C3%A9nom@h', HostSpec('h', None, 'prénom', None)),
+    ('pr%C3%A9nom:p%40ss%3Aw0rd@h', HostSpec('h', None, 'prénom',
+                                             'p@ss:w0rd')),
 ])
 def test_parse_hostspec(hostspec, expected):
     assert BaseBackend.parse_hostspec(hostspec) == expected
+
+
+@pytest.mark.parametrize('hostspec,pod,container,namespace', [
+    ('kubectl://pod', 'pod', None, None),
+    ('kubectl://pod?namespace=n', 'pod', None, 'n'),
+    ('kubectl://pod?container=c&namespace=n', 'pod', 'c', 'n'),
+])
+def test_kubectl_hostspec(hostspec, pod, container, namespace):
+    backend = testinfra.get_host(hostspec).backend
+    assert backend.name == pod
+    assert backend.container == container
+    assert backend.namespace == namespace
 
 
 @pytest.mark.parametrize('arg_string,expected', [