+++ testinfra/backend/base.py
@@ -45,7 +45,30 @@ class CommandResult(object):
         super(CommandResult, self).__init__()
 
     @property
+    def succeeded(self):
+        """Returns whether the command was successful
+
+        >>> host.run("true").succeeded
+        True
+        """
+        return self.exit_status == 0
+
+    @property
+    def failed(self):
+        """Returns whether the command failed
+
+        >>> host.run("false").failed
+        True
+        """
+        return self.exit_status != 0
+
+    @property
     def rc(self):
+        """Gets the returncode of a command
+
+        >>> host.run("true").rc
+        0
+        """
         return self.exit_status
 
     @property
@@ -194,6 +217,11 @@ class BaseBackend(object):
                 user, password = user.split(':', 1)
         if ':' in name:
             name, port = name.split(':', 1)
+        name = testinfra.utils.urlunquote(name)
+        if user is not None:
+            user = testinfra.utils.urlunquote(user)
+        if password is not None:
+            password = testinfra.utils.urlunquote(password)
         return HostSpec(name, port, user, password)
 
     @staticmethod