+++ testinfra/backend/lxc.py
@@ -17,28 +17,16 @@ from __future__ import absolute_import
 from testinfra.backend import base
 
 
-class KubectlBackend(base.BaseBackend):
-    NAME = "kubectl"
+class LxcBackend(base.BaseBackend):
+    NAME = "lxc"
 
     def __init__(self, name, *args, **kwargs):
         self.name = name
-        self.container = kwargs.get('container')
-        self.namespace = kwargs.get('namespace')
-        super(KubectlBackend, self).__init__(self.name, *args, **kwargs)
+        super(LxcBackend, self).__init__(self.name, *args, **kwargs)
 
     def run(self, command, *args, **kwargs):
         cmd = self.get_command(command, *args)
-        # `kubectl exec` does not support specifying the user to run as.
-        # See https://github.com/kubernetes/kubernetes/issues/30656
-        kcmd = 'kubectl '
-        kcmd_args = []
-        if self.namespace is not None:
-            kcmd += '-n %s '
-            kcmd_args.append(self.namespace)
-        if self.container is not None:
-            kcmd += '-c %s '
-            kcmd_args.append(self.container)
-        kcmd += 'exec %s -- /bin/sh -c %s'
-        kcmd_args.extend([self.name, cmd])
-        out = self.run_local(kcmd, *kcmd_args)
+        out = self.run_local("lxc exec %s --mode=non-interactive -- "
+                             "/bin/sh -c %s", self.name, cmd)
+        out.command = self.encode(cmd)
         return out